refactor: extract exception handler
